Improve robustness of engine termination
Change the "NotifyShutdown" to explicitly destroy the engine
and resource manager singleton.

Previously, the global engine object destruction relies
on the automatic static variable destruction on termination.

Destroying the engine and resource manager involves a sequence of
complicated operations, including locking and thread joins. Such
operations are not safe to perform during dll unload, especially on
windows. The outcome is that python hangs on termination.

Also fix a bug in threaded_engine_perdevice where the CPUWorker
could segfault by taking null pointer. This happens when
the engine is created and destroyed immediately. The destructor
resets the threadblock shared_ptr to nullptr, and the CPU worker function
is capturing the shared_ptr rather than the real pointer.